pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lit-anal"
version = "0.1.0"
description = "Literature Analysis Synthesizer"
requires-python = ">=3.11"
dependencies = [
    # UI
    "streamlit>=1.32.0",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    # Zotero
    "pyzotero>=1.5",
    # PDF extraction
    "docling>=2.0",
    # LLM workflow
    "langgraph>=0.2",
    "openai>=1.30",
    # Observability
    "langfuse>=2.0",
    # Config
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    # Utils
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
